refactor(daemon): neutralize the send queue and the interaction-actor type
De-baseclass Slack, part 1: two pieces of platform-neutral machinery lived
under `src/slack/` and were imported by telegram/discord/feishu.

- `SlackSendQueue` -> `PlatformSendQueue` in `src/platforms/send-queue.ts`.
  It never had Slack-specific behavior; only its doc comment claimed Tier-3
  specifics. Rewritten to platform-neutral wording. Nine importers updated,
  no deprecated re-export left behind.
- `InteractionActor` moved from `src/slack/connection.ts` into the Layer-1
  contract file `src/platforms/contract.ts`. The type is self-contained, so
  no import cycle: contract.ts still imports nothing from slack.

Pure moves + renames; zero behavior change. Design docs referencing the old
name/path updated alongside.
