+++ packages/daemon/src/platforms/send-queue.ts
@@ -1,20 +1,19 @@
 /**
- * §9.1 SlackSendQueue: serialize outbound Slack Web API writes (chat.postMessage
- * / chat.update / assistant.threads.setStatus / assistant.threads.setTitle) per connection and space them by a
- * minimum interval so streamed edits don't trip Slack's Tier-3 rate limit
- * (chat.postMessage ~50 rpm). FIFO: preserves the order the converger emits.
+ * PlatformSendQueue: serialize a connection's outbound chat-platform writes and
+ * space them by a minimum interval, so streamed edits don't trip a provider's
+ * per-method rate limit. FIFO: preserves the order the converger emits.
  *
  * The first task runs immediately; each subsequent task waits until at least
  * `minIntervalMs` after the previous one *started*. `enqueue` resolves/rejects
- * with the task's own result, so callers can still await a posted message ts.
+ * with the task's own result, so callers can still await a posted message id.
  *
- * Each task is bounded by `taskTimeoutMs`: if a single call hangs (e.g. a Slack
+ * Each task is bounded by `taskTimeoutMs`: if a single call hangs (e.g. a provider
  * API stall during a socket reconnect), it is abandoned so the queue keeps moving
  * and a best-effort status update can never block real message delivery. The
  * abandoned call's promise still settles in the background; the caller just sees a
  * timeout rejection (which, for a progress/plan post, the daemon treats as "no ts").
  */
-export class SlackSendQueue {
+export class PlatformSendQueue {
   private chain: Promise<unknown> = Promise.resolve()
   // -inf so the very first task never waits, regardless of the clock's origin.
   private lastStart = Number.NEGATIVE_INFINITY
@@ -49,7 +48,7 @@ export class SlackSendQueue {
       const timer = setTimeout(() => {
         if (settled) return
         settled = true
-        reject(new Error(`SlackSendQueue: task exceeded ${this.taskTimeoutMs}ms — abandoned`))
+        reject(new Error(`PlatformSendQueue: task exceeded ${this.taskTimeoutMs}ms — abandoned`))
       }, this.taskTimeoutMs)
       p.then(
         (v) => {
