+++ packages/daemon/test/send-queue.test.ts
@@ -1,5 +1,5 @@
 import { describe, it, expect } from 'vitest'
-import { SlackSendQueue } from '../src/slack/send-queue.js'
+import { PlatformSendQueue } from '../src/platforms/send-queue.js'
 
 /** A fake clock + sleep so spacing is asserted deterministically (no real timers). */
 function fakeClock() {
@@ -15,17 +15,17 @@ function fakeClock() {
   }
 }
 
-describe('SlackSendQueue', () => {
+describe('PlatformSendQueue', () => {
   it('runs the first task immediately (no initial wait) and returns its result', async () => {
     const clk = fakeClock()
-    const q = new SlackSendQueue(1000, clk.now, clk.sleep)
+    const q = new PlatformSendQueue(1000, clk.now, clk.sleep)
     expect(await q.enqueue(async () => 'first')).toBe('first')
     expect(clk.sleeps).toEqual([])
   })
 
   it('preserves FIFO order and spaces subsequent tasks by minIntervalMs', async () => {
     const clk = fakeClock()
-    const q = new SlackSendQueue(100, clk.now, clk.sleep)
+    const q = new PlatformSendQueue(100, clk.now, clk.sleep)
     const order: number[] = []
     const ps = [1, 2, 3].map((n) => q.enqueue(async () => (order.push(n), n)))
     expect(await Promise.all(ps)).toEqual([1, 2, 3])
@@ -35,7 +35,7 @@ describe('SlackSendQueue', () => {
   })
 
   it('a throwing task rejects its own promise but does not break the chain', async () => {
-    const q = new SlackSendQueue(0)
+    const q = new PlatformSendQueue(0)
     const ran: string[] = []
     const p1 = q.enqueue(async () => {
       throw new Error('boom')
@@ -51,7 +51,7 @@ describe('SlackSendQueue', () => {
 
   it('abandons a hung task after the per-task timeout so the queue keeps moving', async () => {
     // real timer for the timeout; fake clock only for spacing (minInterval 0)
-    const q = new SlackSendQueue(
+    const q = new PlatformSendQueue(
       0,
       () => 0,
       async () => {},
